@@ -1,11 +1,10 @@
 import Artiste from 'Artiste'
-import Helmet from 'react-helmet'
 import React from 'react'
-import moment from 'moment'
+import Helmet from 'react-helmet'
 
 import InProgressChecker from 'InProgressChecker'
 
-import { vendredi as groupes } from '../data/groupes.json'
+import prog from '../data/groupes.json'
 
 export default class Vendredi extends React.Component {
   render() {
@@ -20,7 +19,7 @@ export default class Vendredi extends React.Component {
             Entrées : participation libre (Chacun donne ce qu’il veut ou ce
             qu’il peut)
           </p>
-          {groupes.map((groupe, index) => (
+          {prog.vendredi.map((groupe, index) => (
             <Artiste groupe={groupe} key={index} />
           ))}
         </InProgressChecker>
